# tests/fixtures/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookCreate": {
        "description": "Request body for POST /webhooks.\n\nReference: GitHub Webhooks; Stripe Webhooks; DESIGN.md \u00a710.25 (v0.30.0).",
        "properties": {
          "events": {
            "items": {
              "type": "string"
            },
            "title": "Events",
            "type": "array"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "events"
        ],
        "title": "WebhookCreate",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/webhooks": {
      "get": {
        "description": "Return all registered webhooks with delivery statistics.\n\nEach entry contains:\n- ``id``: webhook UUID\n- ``url``: target URL\n- ``events``: subscribed event names\n- ``created_at``: Unix timestamp of registration\n- ``delivery_count``: total delivery attempts\n- ``failure_count``: total failed attempts\n\nDesign reference: DESIGN.md \u00a710.25 (v0.30.0).",
        "operationId": "list_webhooks_webhooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response List Webhooks Webhooks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List all registered webhooks"
      },
      "post": {
        "description": "Register a new outbound webhook.\n\nWhen a subscribed event fires, the orchestrator POSTs a JSON payload to\nthe registered URL.  An optional HMAC-SHA256 signature is included in\nthe ``X-Signature-SHA256`` header when ``secret`` is supplied.\n\nValid event names:\n``task_complete``, ``task_failed``, ``task_retrying``, ``task_cancelled``,\n``task_dependency_failed``, ``task_waiting``, ``agent_status``,\n``workflow_complete``, ``workflow_failed``, ``workflow_cancelled``, ``*``\n(wildcard \u2014 receive all events).\n\nReturns: ``{id, url, events, created_at}``\n\nDesign reference: GitHub Webhooks; Stripe Webhooks; RFC 2104 HMAC;\nZalando RESTful API Guidelines \u00a7webhook; DESIGN.md \u00a710.25 (v0.30.0).",
        "operationId": "create_webhook_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Webhook Webhooks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a new webhook"
      }
    },
    "/webhooks/{webhook_id}": {
      "delete": {
        "description": "Remove a registered webhook by ID.\n\nReturns 404 if the webhook ID is unknown.\n\nDesign reference: DESIGN.md \u00a710.25 (v0.30.0).",
        "operationId": "delete_webhook_webhooks__webhook_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "webhook_id",
            "required": true,
            "schema": {
              "title": "Webhook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Webhook Webhooks  Webhook Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a webhook"
      }
    },
...
  }
}